refactor websocket / power usage
