+++ classes/PowerUsage/index.js
@@ -1,0 +1,7 @@
+const DeviceRaw = require('./DeviceRaw');
+const CurrentMonth = require('./CurrentMonth');
+
+module.exports = {
+  DeviceRaw,
+  CurrentMonth,
+};
