--- lib/powerUsage/index.js
@@ -1,7 +1,0 @@
-const deviceRawPowerUsage = require('./deviceRawPowerUsage');
-const currentMonthPowerUsage = require('./currentMonthPowerUsage');
-
-module.exports = {
-  deviceRawPowerUsage,
-  currentMonthPowerUsage,
-};
